pyproject.toml
[tool.poetry]
name = "logme"
version = "0.1.0"
description = ""
authors = ["Jonathan MERCIER <jonathan.mercier@microbiome.studio>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
email-validator = "^1.2"
pydantic = "^1.9"
PyYAML = "^6.0"
sqlalchemy = "^1.4"
pyjwt= "^2.4"
jwcrypto = "^1.3"
passlib = "^1.7"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.3"
asyncpg = "^0.26.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
black = "^22.3.0"
mypy = "^0.971"

[tool.poetry.scripts]
msio_logme = "msio.logme.__main__:main"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[flake8]
max-doc-length = 72

[tool.black]
line-length = 75
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3
py_version = 310
line_length = 75

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
